refactor : 테스트 코드 접근제어자 통일, 메소드 명 수정
