+++ src/test/java/calculator/StringAddCalculatorTest.java
@@ -10,12 +10,13 @@ import org.junit.jupiter.params.provider.ValueSource;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.AssertionsForClassTypes.assertThatThrownBy;
 
+@SuppressWarnings("NonAsciiCharacters")
 class StringAddCalculatorTest {
 
     @DisplayName("null 또는 빈 문자일 경우 0을 반환")
     @ParameterizedTest
     @NullAndEmptySource
-    public void splitAndSum_null_또는_빈문자(String input) {
+    void splitAndSum_null_또는_빈문자(String input) {
         int result = StringAddCalculator.splitAndSum(input);
 
         assertThat(result).isEqualTo(0);
@@ -58,7 +59,7 @@ class StringAddCalculatorTest {
     @DisplayName("음수 또는 숫자 이외의 값을 입력할 경우 RuntimeException 예외를 throw")
     @ParameterizedTest
     @ValueSource(strings = {"-1,2,3", "a,b,c", "//;\n1;2;a"})
-    public void splitAndSum_negative() {
+    void splitAndSum_negative() {
         assertThatThrownBy(() -> StringAddCalculator.splitAndSum("-1,2,3"))
                 .isInstanceOf(RuntimeException.class).withFailMessage("입력 값이 음수이거나 숫자가 아닙니다.");
     }
